requirements.txt
boto3==1.33.12
Flask==3.0.0
psycopg2-binary==2.9.9
requests==2.25.1